Update Readme.md
A more detailed description regarding the Unit (activity) for purchased enzymes was added. Now the note is included that the activities determined by the supplier might differ form the ones calculated by your own. furthermore, a reference should be given for the biocatalyst so that the same charge can be reordered if necessary and available.