Update Readme.md
A description of the differences between immobilized enzymes in a packed-bed or fluidized bed versus soluble enzymes in a liquid phase is now provided here. When selecting the reactor type, it is essential to consider the state in which you intend to apply the enzymes (immobilized or soluble). In addition, concerning the 'Mixing' sub-category, it was crucial to include attributes for cases where a stir bar is utilized in the experiment.